ask before delete project

+++ src/views/Delete.vue
@@ -20,25 +20,34 @@ export default {
             const uuid = route.params.uuid;
             const token = sessionStorage.getItem("mypt_token");
 
-            axios({
-                method: "DELETE",
-                url: config.api.host + `/manage/${uuid}`,
-                headers: {
-                    Authorization: `Bearer ${token}`,
-                }
-            }).then((e) => {
-                console.log(e);
-                router.push({ name: "AboutMe" });
-            }).catch((e) => {
-                const code = e.response.status;
-                const data = e.response.data;
+            if(
+                confirm(
+                    "* 프로젝트를 삭제할까요?"
+                )
+            ){
+                axios({
+                    method: "DELETE",
+                    url: config.api.host + `/manage/${uuid}`,
+                    headers: {
+                        Authorization: `Bearer ${token}`,
+                    }
+                }).then((e) => {
+                    alert(e.data.message);
+                    router.push({ name: "AboutMe" });
+                }).catch((e) => {
+                    const code = e.response.status;
+                    const data = e.response.data;
 
-                if(code == 404){
-                    alert(data.error.message);
-                } else {
-                    alert(code + ": " + data.message);
-                }
-            });
+                    if(code == 404){
+                        alert(data.error.message);
+                    } else {
+                        alert(code + ": " + data.message);
+                    }
+                });
+            } else {
+                alert("프로젝트 삭제가 취소되었습니다.");
+                router.push({ name: "Project", params:{ uuid: uuid } });
+            }
         }
     }
 }
